LPFG-975: Moved NotificationService to /src/lib/service
